Fix (runtime) ImGui warning about duplicate IDs
Newer Dear ImGui version now checks for this. When hovering over widgest
with duplicate ID, you now get a popup with a warnings and all widgets
get a red border.

In this specific case the duplicate IDs (the ID was 5 times an empty
string) didn't cause any harm. But it's good to avoid anyway ... it
would have caused very weird behavior if we wanted to extend the
functionality in any way.