Issue#1780: Toolbar buttons have inconsistent sizes and shapes
https://github.com/openMSX/openMSX/issues/1780

I never noticed this before, but after reading the ticket I do see it
and now it annoys me :-)

The underlying technical reason is that these icons are actually glyphs
in a (custom) font. Though these glyphs have slightly different sizes.
And I didn't find a way to create a ImGui::ImageButton() where the size
of the button is not directly controlled by the size of the image. E.g.
center a smaller image in a larger button.

As a workaround, I now calculate the largest size (in X and Y direction)
of all icons, and then stretch (all) icons to that common size. Though
this makes (some) icons (slightly) blurry. It's not ideal, but at least
IMHO it's already an improvement. For the real solution someone (not me)
will have to create new glyphs.