Refactor gl::PixelBuffer, replace width/height with size
Before PixelBuffer represented a buffer of 'width x height' number of
pixels, so a buffer with a particular shape. Though in fact (at the
openGL level) this is just a 1D buffer. For example a buffer of 2000
pixels can hold an image of 40 x 50 pixels, but just as well an image of
20 x 100 pixels. And also smaller images like 20 x 20.

In a future patch (to generalize the hq-scaler to more than 320x240) we
want to reuse the same PixelBuffer for multiple shapes. So instead of
allocate one for each shape, we allocate only one that can hold the
largest shape.

This patch also changes the 'mapped()' method to return a 'std::span<T>'
instead of a raw-pointer 'T*'. This makes the API more descriptive (and
simplifies some caller code). And it also makes it safer: we now have
(optional) out-of-bounds checks.

The 'getOffset(x, y)' method is removed. It no longer makes sense now
that the buffer no longer has a specific shape. Also for cache
efficiency it's better to as much as possible reuse the same region in
the buffer, so typically the start of the buffer, and that's already
available via the above `mapped()` method.