Make opengl scalers work with both 320 and 640 width screens PR#1761
This commit is heavily inspired by
    ttps://github.com/openMSX/openMSX/pull/1761
Thanks to a lot to balyuf, both for reporting and creating the initial
solution.

Summary (more details in the above link):
The 'hq', 'hqlite' and 'ScaleNx' used to work in 'screen 7' with the old
'SDL' renderer (work as in: they don't fallback to the default scaler).
They never worked with the 'SDLGL-PP' renderer. We did remove the 'SDL'
renderer, so in a way this can be seen as a regression.

The 'SDL' version of the hq{lite} scaler had special code for scaling
the X-dimension by half the amount of the Y-dimension (and that's what
happens in 'screen 7' non-interlaced modes): The effect is as-if the
full hq-scaler output is calculated, and then take the average of all
pixel-pairs. Instead this implementation only calculates every other
output pixel. We may still want to change this in the future.

The main differences between this commit and the approach taken in
PR#1761 are:
* This also works for gfx9000 modes, and then specifically B7 (1024
  pixels wide).
* Instead of having separate buffers/textures for width=320 and
  width=640 (and after the previous point also width=1280), we only have
  one buffer/texture with the maximal size (and then maybe only partly
  use it).