Issue#1780: Toolbar buttons have inconsistent sizes and shapes
The suggestion in this commit is to change the ImageButton to a Button and convert the glyph symbol to a null terminated character. This makes it easier to align and center the symbol in the button without undesired scaling. See issue @1780, PR #459f221f56328ba0bbd97bd988edef6d825dffd5