[1/11] List available savestates with C++ code only, like with replays.
Preparation to make it easier to show more information about the states in the
menu.

Also: define the savestate file extension and folder in one place, at least
within this file. (There's no obvious place in our non-GUI code to put it...)

This change also changes the behaviour: if there are no savestates, the menu
item is not disabled anymore. In a later commit we will restore something
similar, but easier to understand: we explicitly tell no savestates were found.