[9/11] Make the loadstate/loadreplay tables narrower.
They were way too wide.

Accomplished by showing only the date or only the time (when it's made today).
The full date/time is shown in a tooltip. Also made a tooltip with the full
name, so we can make the name column narrower without worrying about long names
not being readable.

To avoid double calculations with the time/date functions, we already convert
to std::time whilst scanning for files.