Remove 'hqlite' scaler algorithm
It was added as a more light-weigth version of the 'hq' algorithm
(instead of checking if colors are similar it checks if colors are
equal). 20 years ago that was important, but nowadays not so much. Or
better said: the extra quality is more important than the speed.

Though the main motivation for removing 'hqlite' is to simplify our code
base (just a little).