[pre-2] Only use UNaligned SSE load/store intrinsics
For example replace:
    _mm_load_si128()
with
    _mm_loadu_si128()
The former requires a 128-bit aligned pointer, the latter doesn't have
this requirement.

When SSE was first introduced, the former instruction would execute
faster. But on modern CPUs both instructions execute:
* Just as fast when the given pointer is actually aligned.
* Just as fast when the pointer is not fully aligned, but does not cross
  a cache-line boundary (I mean the 128-bit element starting at the
  given pointer).
* If the element does cross a cacheline or a page, then there might be a
  performance penality (for aligned accesses such crossings can never
  happen).

So nowadays there's much less need to use aligned load/stores. And
using the unaligned versions does allow some simplifications in the rest
of the code (see next patches). Nevertheless compiler still try to align
large buffers. So in practice the unaligned load/store instruction will
still operate on aligned address (IOW it's still likely but no longer
100% guaranteed).

--

This is a preparation to be able to get rid of
   VLA_SSE_ALIGNED
This alignment was useful in the past, but much less so nowadays.