Resolve sonar cloud remark about explicit constructor/conversion-operator
Before we added a comment /*implicit*/ to document the missing
"explicit" was intended. Since c++20 you can 'calculate' whether or not
the function should be explicit. And with this "explicit(false)" has
become a defacto-standard way to mean "implicit". It means exactly the
same as no annotation at all, but it does make the intent clear: it's
intended like this, we did not simply forget about it.