Extend MemBuffer with size
Before MemBuffer internally only kept a pointer to the allocated memory
block. This patch extends that with a size member variable.

This makes MemBuffer objects a bit more heavy (sizeof(MemBuffer)
increases from 8 to 16). Though often (not always):
* We manually kept track of the size in some other way (future patches in
  the series will then eliminate those).
* The MemBuffer object is a local variable, and then the compiler can
  usually eliminate the extra size member (if indeed unused).

This size member allows to add (optional) bounds-checking in MemBuffer.
And it allows to (correctly) convert MemBuffer into a std::span, this
also allows for better bounds and buffer-overflow checks.

More in detail, this patch extends MemBuffer with these convenience
methods:
* size() and empty()
* begin() and end()
* front() and back()
* operator[] with (optional) bounds-check
* conversion to span, first(), subspan()

This brings MemBuffer a bit closer to std::vector. The remaining
differences are:
* std::vector is still a bit more heavy weight. Typically sizeof(vector)
  == 24, sizeof(MemBuffer) == 16.
* MemBuffer keeps the initial content uninitialized. std::vector always
  initializes, even if the buffer will soon after be overwritten (e.g.
  by reading from a file).
* Because of this it's only allowed to use MemBuffer with trivially
  copyable types (copy is equivalent with memcpy). This patch adds
  such a check.
* On destruction MemBuffer just frees the memory block. std::vector will
  first properly destruct all elements it contains. But it's possible
  the destructor of those elements is trivial (does nothing), then of
  course std::vector also doesn't need to do anything extra.
* Because of this it's only allowed to use MemBuffer with trivially
  destructible types. This patch adds a check for that.
So the remaining difference are still sufficient to warrant the
existence of MemBuffer (we cannot just replace it with std::vector and
keep the same performance).

This change in MemBuffer allows for various simplifications. But that's
left for (multiple) follow-up patches.