Small tweak on c9533bc1ccd1784
That commit adds the text "page=n" (with 'n' a digit 0-3). Because 'n'
can vary we want to show it in a monospace font, otherwise the text that
follows on the same line can shift around.

I discussed this with David (turbor). Initially we thought it was fine
to show the full text (page=n) with the same monospace font. Though now
when I actually see the result, you can notice the difference in font
between "page=" and the other text (non-digits) on the line.

It's a detail, but let's fix it anyway.