Refactor interaction between rapidsax and XMLElement
Triggered while investigating the clang-tidy suggestion
'bugprone-suspicious-stringview-data-usage'.

Our XML parser (rapidsax) calls various handlers with std::string_view
parameters, e.g. to pass the name of an XML-tag. Optionally we can
request to zero-terminate those strings (this happens in-place in the
memory buffer containing the XML file).

For example in XMLDocument::load() we request such zero-termination. It
has a callback function
    void start(std::string_view tagName)
That callback uses 'tagName.data()' and then (indirectly) passes that to
the XMLElement constructor which takes a 'const char*' parameter. This
parameter is expected to point to a zero-terminated string.

In general that's a suspicious operation: in general
std::string_view::data() may not point to zero-terminated data. And
that's why the clang-tidy gave a warning on it.

In this specific case we know that the std::string_view points to a
zero-terminated string. Thus all is fine. But it's always nice to
express (and verify) such guarantees/constraints via the c++ type-system
(so that you get a compile-error when you violate the constraint).
That's possible by changing the type from 'std::string_view' to
'zstring_view'. (Note: we only do this when we do request
zero-terminated strings, we don't always require that, and then we
continue using 'std::string_view').

However, while implementing this change (string_view -> zstring_view), I
noticed we do actually violate this constraint: we do pass
NON-zero-terminated strings to the start() callback!!! Or at least we
temporarily violate it. Only some time after the callback returns we do
add the zero-terminator.

So in the end it worked fine, but nevertheless it was very brittle code.
So this commit also ensures that we zero-terminate BEFORE calling the
callback functions. This has a (small) cost in rapidsax::parseElement(),
but that's an acceptable trade off for the cleaner semantics / safety
guarantees.