Apply a few of clang-tidy 'bugprone-implicit-widening-of-multiplication-result' suggestions
This clang-tidy warning triggers e.g. when you multiply 2 32-bit values
together (the result is still 32-bit, thus it might overflow) and then
convert the result to 64-bit (which could hold the result without
overflow).

The 'fix' is to first convert (at least one of) the operands to 64-bit
and then perform a 64-bit multiplication(*).
 (*) So in C++ you do a '64-bit x 64-bit -> 64-bit', but the compiler
     may turn that into '32-bit x 32-bit -> 64-bit' if that's more
     efficient on the target CPUi (e.g. on x86).

Even if in practice there's no risk of overflow it may still be
beneficial. Consider this example:
    myVector.reserve(myUint32 * 2);
Suppose _we_ know myUint32 is always small, so the multiplication will
never overflow. But _the_compiler_ may not know this (and/or take this
into account). So it still has to generate code as-if overflow _could_
happen.

On a 64-bit CPU the variable 'myUint32' is likely stored in a 64-bit
register. So the multiplication by 2 (or shifting) does not
(necessarily) automatically truncate this to 32-bit. So the compiler
does need to insert an explicit instruction to zero the 32 upper bits
before it can be passed to the reserve() function (which does take a
64-bit input).

If instead we would have written this code:
    myVector.reserve(myUint32 * size_t(2));
We both avoid the potential overflow, and we possibly save one
instruction.

So the 'bugprone-implicit-widening-of-multiplication-result' warnings
seems like a good idea. Unfortunately in its current form they produce
_way_ too many false positives.

For example it also triggers on code like this:
    std::array<int, 10 * 2> myArray;
This indeed does an implicit widening of a 32-bit multiplication to
64-bit. But in this case that perfectly fine. We're very sure there is
no overflow here ;)

So unfortunately we had to ignore most of these clang-tidy warnings.
This patch only applies the very obvious ones. (And it likely misses
some, the signal-to-noise ration is just too high).