Replace horizontal/vertical selector in ImGuiOsdIcons
Before this patch you could select between a horizontal or a vertical
layout for the OSD icons. This patch removes that selector and just
fills the available window space row-by-row. You obtain a vertical
layout by making the window very narrow (so that only 1 icon fits per
row). I believe this idea was once suggested by 'Vampier'.

This does make a tradeoff:
* Before it was possible to control the spacing between the icons (by
  making the window wider/taller than needed). That's no longer
  possible.
* But now it's possible to arrange the icons in more than one row.

So this is an experiment. Let's see how it goes.