Fix (=remove) assertion about alignment
A little while ago we no longer strictly align (some) buffer to 16-byte
boundaries (for x86 SSE instructions). See commit fc04f341a3f92349 and
follow-ups for the rationale.

Though we still had a few asserts in our code that do check for this
stricter alignment. It turned out that this assert still passed when
compiled with gcc, but not with clang. Apparently gcc figures out it's
still a good idea to keep large buffers 16-byte aligned (or it could
just be a coincident).

Anyway, removing the assert fixes the problem. And we could cleanup some
more alignment-check related code as well.