Tweak BreakPointBase::getCondition() and getCommand()
Before we had two versions of each:
* getCondition() and getConditionObj()
* getCommand()   and getCommandObj()
The former returns a string, the latter a TclObject.

Though for all callers that query the string it was actually more
efficient to instead query the TclObject (e.g. because afterwards the
string is converted back to a TclObject). This is probably a left-over
from when TclObject was less flexible.

So in this patch:
* We change all callers to use the 'obj' variant.
* We remove the non-obj variant.
* We rename getConditionObj() to getCondition() (and same for getCommand()).