Allow to specify an id-attribute for the initial flash content in the config file
By including a section like
    <rom>
      <filename>...</filename>
      <sha1>...</sha1>
    </rom>
in the config file of a MSX device it's possible to use that rom as the
initial content for the flash rom. This is inly for the first time that
device is started. The next time it's started it uses the last, possibly
re-flashed, content.

Though this is ambiguous when the MSX device contains multiple roms. A
concrete example is the DalSoRiR2 device (which we'll add soon). It
contains both a flash rom (for the MSX-AUDIO BIOS) and the OPL4
sample-rom.

This patch allows to (optionally) specify an "id-string" for the flash
content. Then only the config section with corresponding id will be
taken.
    <rom id="...">
      <filename>...</filename>
      <sha1>...</sha1>
    </rom>

Or when no rom-tag with matching id is present in the config file, then
no initial content is loaded (initial flash content is all 0xff).