Move YMF278B specific knowledge out of MSXMoonSound class
The OPL4 (YMF278B) chip has two main parts:
* The FM-part, this is implemented in the YMF262 class.
* The wave-part, this is implemented in the YMF278 class.
Both parts are mostly, but not fully independent.

Before this patch a lot of the glue-logic that ties both parts together
was implemented in the MSXMoonSound class. This patch moves that logic
to a new class YMF278B.

The main motivation for this is that we want to add a new MSX device
DalSoRiR2 which also contains a OPL4 chip. And we don't want to
duplicate all that glue-logic.