Made YMF278 independent from MoonSound-specific memory layout
The MoonSound cartridge has a 2MB ROM and a variable amount of RAM
connected to the OPL4 chip (in a specific way, see code for details).
Currently that knowledge is hardcoded in the YMF278 class.

We want to add the DalSoRiR2 device. That one also uses a OPL4 chip, but
it has the memory connected in a different way.

This patch makes the YMF278 class independent from the specific memory
layout. More specifically devices using the OPL4 chip must pass a
callback that is responsible for creating the specific memory layout.
This must be callback, and not some static setup because the YMF278 can
dynamically switch between two memory modes (see code for details).