Fix debug build / add optional_fixed_span utility
In the YMF278 code we divide the 4MB sample address space into 32 chunks
of 128kB. For each of these chunks we store a pointer to the
corresponding mapped memory (ROM or RAM, possibly mirrored) or nullptr
in case that chunk is unmapped.

Though instead of storing a raw pointer we store a
    std::span<const uint8_t, 128*1024>
indicating it points to a 128kB chunk (a raw pointer does not indicate
how large the pointed to memory block is). And internally a fixed size
span also stores just a pointer. So far so good.

The problem is when we want to indicate an unmapped memory region. An
invariant of std::span<T, N> is that it must _always_ points to a valid
range of N elements. But we can try to 'trick' the compiler via an
initialization like:
    std::span<T, N>{static_cast<T*>(nullptr), N};
That compiles, and actually runs fine. And we can check the invalid
state via something like:
    data() == nullptr
However the debug-STL-mode in gcc verifies several invariants, including
the one that a 'span<T, N>' must point to a valid range. And thus that
triggers an error on the above initialization.

To fix this, and keep the same efficiency, we introduced the
'optional_fixed_span<T, N>' class:
* It can only be constructed from
  - a valid std::span<T, N>
  - or to the empty state
* Internally it stores a single pointer.
* And the only operation on this class is: convert it to a
   std::optional<std::span<T, N>>
* This enforces the check for nullptr (via std::nullopt).
* And it retains the information that it points to N elements.

I (manually) verified that the generated code is as efficient as using
plain raw pointers.