[1/2] refactor WatchIO
Remove 'motherBoard' member variable. Instead pass it as a parameter to
the methods that need it.

This is a preparation to be able to change the properties of an existing
WatchPoint (without having to delete the object and create another one).
For this we need to be able to unregister and then re-register.