[2/2] merge WatchIO into WatchPoint
We want to be able to change the properties of a WatchPoint without
having to delete and re-create the object, including the type of the
watchpoint ({read,write}-{memory,IO}). For that to be possible we need
the same class to represent all these types.

Before this patch IO-watchpoints were represented via WatchIO and
memory-watchpoints via WatchPoint (which is a baseclass of WatchIO).
This class merges both into a single class.