Don't keep list of breakpoints in sorted order
Before this patch we kept the list of breakpoints sorted on address.
That allowed to use binary search to see if the current-pc is hit by a
breakpoint. With this patch that becomes a linear search.

This may seem like a performance degradation, and when there are many
breakpoints (+10(?), or +20(?), I didn't actually measure) it indeed
will be slower. Though typically there won't be that many breakpoints,
and then a linear search will be just as fast (or even faster if there
are only a few). Note that the most important (and most common) case of
no breakpoints at all is already taking a different code path (and there
we don't do any breakpoint-check at all in the inner loop).

If performance ever does become an issue (in case of many breakpoints)
then we can still optimize it via a std::bitset. Then the check will be
very fast, and independent of the number of breakpoints (O(1)).

But the motivation for this change is not related to performance at all.
Instead, in some later commit, we want the GUI to use the openMSX
breakpoint list directly. And then we want to be able to reorder the
list to (re-)sort the table.