Move 'prefix' into BreakPoint, WatchPoint, DebugCondition
We had a lot of places in the code that duplicated the knowledge of how
to represent a breakpoint (or watchpoint or debug-condition) ID. So
stuff like:
    strCat("bp#", bp.getId());
This patch replaces that with a new getIdStr() method.

Detail: some code used strCat() (resulting in a std::string), most code
used tmpStrCat() (resulting in a TemporaryString). In general the latter
can be more efficient because it (more often) avoids temporary memory
allocation. Though in this specific case the resulting strings are small
enough so that (usually) also std::string can avoid the allocation.