Keep formatting of {Break,Watch}Point address
Before this patch the (begin or end) address of a breakpoint/watchpoint
was stored as an integer. And then e.g. for the list command we would
format it as a 4-digit hex number. In the GUI we would already
separately keep track of the original address string as entered by the
user. This patch now also stores the original string in the breakpoint
object itself (and then later patches will remove it from the GUI side).

To convert the string to a number we evaluate it as a Tcl expression.
That means that symbols are supported via the syntax:
    $sym(mySymbolName)
Via the GUI we still allow the simplified syntax of using plain symbol
names. But internally that is then transformed into the above Tcl
expression.

Note that for performance reasons those Tcl expressions are not
constantly being re-evaluated. E.g. an expression like '[reg HL]' to set
a breakpoint on the address where the HL register points to will not
automatically move the breakpoint whenever the HL register changes. It
is _intentionally_ left unspecified when the expressions are being
re-evaluated.