Fix for issue #1756: Menu bar doesn't (reliably) appear on full screen in Windows, in contrast to Linux
Do not use SDL_SetRelativeMouseMode which is meant to pin the absolute mouse, causing the menubar to be unreachable in Windows.
It seems the pinning is done differently or not at all in Linux, hence not showing the issue.
The alternative SDL_SetWindowGrab does not pin the mouse, so that one is now used.
However this causes the cessation of mouse events when reaching the edges, which is why pinning was used in the pinning method.
To fix that issue with this non-pinning method, the mouse is now warped away from the edges when it is near one.

Originaly it was thought that we could simply do without grabbing the mouse, but that caused issues with multiple monitors,
or with explicitly enabling the grab option in windowed mode.