Add -enabled property to breakpoints/watchpoints/debug-conditions
(Part of) the original goal of this patch series was to being able to
set disabled breakpoints from withing a Tcl script:
    https://github.com/openMSX/openMSX/issues/1806
This commit adds such the property '-enabled'. And that allows to do
things like:
    debug breakpoint create -address 0x123 -enabled false
    debug watchpoint configure wp#10 -enabled true

This commit only adds the property in the openMSX core. The GUI doesn't
use it yet.