Use -enabled property on the GUI side
The previous patch introduced the '-enabled' property on breakpoints,
watchpoints and debug-conditions. This patch uses that on the GUI side.

More in detail: before this patch in the openMSX core there only existed
enabled breakpoints (and watchpoints etc). On the GUI side it was useful
to be able to temporarily disable a breakpoint. This was done by
removing it from openMSX and then later, when it's enabled again,
re-adding it. And this was one of the main reasons why the GUI kept it's
own list of breakpoints (largely duplicate from the list in openMSX).
Disabled items in the GUI-list were indicated with a negative ID, items
that also exist on the openMSX side had a positive ID.

This patch removes the concept of 'negative IDs'. Now all elements
on the GUI side always also exist on the openMSX side. For now the list
still exists on the GUI side (and we have code to keep both in sync).
But later patches in this series will fully remove the GUI list.