disassembly view: work on openMSX core breakpoints rather than gui-breakpoints
Next to the main ImGui 'Breakpoints' window, also the 'Disassembly'
window operates on breakpoints. Before this worked with GUI-breakpoints,
e.g. because it also needed to work on disabled breakpoints. But since
the last few commits this is no longer needed. And now the disassembly
view directly works with core openMSX breakpoints.