ImGuiBreakPoints: remove commented out load/save code
At some point in the past, when you exited and then restarted openMSX,
it would restore your breakpoints, watchpoints and debug-conditions.
Though doing that by default was a bad idea, and the feature was
disabled.

It may be a good idea to re-introduce this feature in the future, but
then as an optional feature, and part of a larger session management
feature. For that reason we didn't fully remove the feature from the
code, but only commented it out.

But the previous/next commits in this series totally change the format
and the location of this information (e.g. moved from GUI to core
openMSX). So there's not much point in keeping the old commented out
code. This patch removes it.