Remove state from ImGuiBreakPoints
All previous commits in this series were largly preparations for this
commit. This commit fully removes all state from ImGuiBreakPoints (well,
the only remaining state is whether the window is visible or not).

This allows to remove a lot of code: all code that was responsible for
keeping the state in sync between openMSX core and the GUI.

This is still a rather large patch, I couldn't find an easy way to split
it into smaller chunks.

This basically finishes issue 1806:
    https://github.com/openMSX/openMSX/issues/1806
the remaining commits in this series still make some cleanups.