Move ID-stuff from BreakPoint/WatchPoint/DebugCondition to BreakPointBase
The implementation was identical for all three subclasses. The main
reason it wasn't moved before was because of a static counter 'lastId'
in all subclasses. And this needs to be a separate counter for all
classes (moving it to the base class would make it a shared counter).

The 'solution' is to turn the base class into a CRTP class. That way
it's a template with 3 separate instantiations, thus also separate
counters.

This became possible after earlier changes in this series: we no longer
need a common (non-templatized) base class.

And also for ProbeBreakPoint ... that also allowed to do the
prefix-stuff from an earlier patch in this series.