ImGuiConsole: allow to cancel multi-line command with ESC
Tcl commands can span multiple lines. E.g. when a line has unbalanced
opening curly braces we need extra lines to obtain a full Tcl command.
More technically we use 'Tcl_CommandComplete()' to check if a command is
complete or whether we need extra lines. Visually the console prompt
changes from ">" to "|" to indicate we expect the continuation of the
previous line, rather than a new command line.

Usually this works fine. And it's similar to how other shells (e.g.
bash) handle this. Though if you made a typo, and accidentally typed an
incomplete command (e.g. unbalanced braces or quotes) then it's not so
easy to get back into the 'normal' state again.

E.g. in 'bash' you can cancel a multi-line input via 'CTRL-C'. And
before we migrated the console to ImGui this also worked in openMSX.
Though in ImGui 'CTRL-C' is already used for "copy". Instead we use ESC
to clear the current line. This commit extends that to also cancel an
in-progress multi-line command.