Per-channel balance other than {-100,0,+100}, use in Yamanooto
We could already specify the balance for individual channels. E.g. the
Pioneer PX-7 machine set the PSG channels:
  channel 1: center
  channel 2: left
  channel 3: right
This was done via a section in the config file:
  <sound>
    ...
    <balance channel="1">   0</balance>
    <balance channel="2">-100</balance>
    <balance channel="3"> 100</balance>
  </sound>
Before this patch the value for the balance setting was only allowed to
be:
  -100: fully left
     0: center
   100: fully right
This patch generalizes that to all values between -100 and 100.

This patch also uses that feature for the Yamanooto SCC device. Though
instead of via the config file, it sets per-channel balance settings
directly from the C++ code:
  channel 1:  50% left, 50% right      (values specified in the Yamanooto manual)
  channel 2:  66% left, 33% right
  channel 3:  33% left, 66% right
  channel 4:  66% left, 33% right
  channel 5:  33% left, 66% right
Though I have to admt that personally I don't really notice this pseudo
stereo effect when playing Konami games. In my tests I had to use more
extreme values (like 20%/80%) to verify that the implementation was
actually working as intended.