ImGuiDebugger: allow to open multiple disassembly views
This can be useful if you want to see multiple assembly routines at the
same time (e.g. to compare them).

By default the first view will scroll to the PC on break. All other
views will not (by default). But this behavior is now configurable
(right click context menu).

When selecting 'Show in disassembly' in the breakpoints or symbols
window, this only has effect on the first disassembly view.