ImGuiBreakPoints: also recognize '$::sym(name)'
So both:
  $sym(name)
  $::sym(name)

In a Tcl proc the extra '::' are needed to refer to the global variable
'sym' (rather than a local variable in the proc). But the breakpoint
address expression is always evaluated in the global context, so there
the extra '::' is not needed. It is allowed, and for consistency I can
see why you'd want to write it like that. So now the GUI recognized both
formats.

Note: this is not related to evaluating the expression, it's purely the
representation in the GUI: the GUI recognizes some specific Tcl
expression patterns and then replaces those with a simpler notation. And
similarly if you type that simpler notation in the GUI, it's internally
replaced with the full Tcl expression.