Further tweak the status bar
Use CheckBox() instead of MenuItem(): this keeps the menu open after
(un)checking an item.

Also add these as a right-click context menu in the status bar itself.
So we have both now. That allows to more easily compare both
alternatives. And maybe we can just keep both? The code is actually
shared (so no duplication).