Don't print info message about screenshot when making a savestate
As this might be confusing.

The screenshot command was partly implemented in c++ (most stuff) and
partly in Tcl (this add the -no-sprite functionality, doing this in Tcl
is easier: we disable sprite rendering, emulate for one frame and then
restore sprite rendering). This was achieved by renaming the c++
"screenshot" command (I mean really via the Tcl 'rename' functionality)
and then defining a Tcl proc with the name "screenshot" (internally that
Tcl proc calls the c++ version). But this effectively hides the original
c++ version.

Instead of renaming the c++ version, this patch gives the c++ version a
different name to start with: "openmsx::internal_screenshot". This keeps
the command available for other users. But as the name suggests it
should only be called from other internal openMSX functions.

Apart from the (re)naming stuff, this patch moves printing the info
message "Screen saved to $filename" from c++ to Tcl. It also moves the
help text and tab-completion from c++ to Tcl (the c++ version of help
and tab-completion already anticipated the extensions that would happen
in the Tcl wrapper, that was a bit weird).

But the whole motivation for this commit is the following change: the
"savestate" command (a Tcl proc) now uses the
"openmsx::internal_screenshot" command instead of just "screenshot".
That's fine, it doesn't need the "-no-sprites" functionality. But more
important: this now no longer prints an info message about screenshots.