Fix (future) clang compile error
Not yet a problem now, but it does become a problem when we pass
-std=c++23 to clang. And clang is right, compilers are allowed to reject
this code.

But the fix is very easy, just swap the order of two classes (so that
when compiling 'unique_ptr<MSXWatchIODevice>' inside 'WatchPoint', the
full definition of 'MSXWatchIODevice' is already availble (not only a
forward declaration).