Use c++23 std::unreachable()
We still keep our UNREACHABLE macro, because in debug mode we define it
as "assert(false)" (to verify that it is indeed unreachable), and in a
release build we define it like "std::unreachable()" (to let the
compiler optimize on the assumption that this location indeed cannot be
reached).

So this patch only simplifies the definition of the UNREACHABLE macro.