Fix msvc compile error: change '/std:c++23' into '/std:c++latest'
Apparently '/std:c++23' is not supported yet (possibly because c++23
support is not yet complete?). BUT an unknown option doesn't give an
error, only a warning. And the code continued to compile fine without
the previous '/std:c++20' option (maybe because c++20 mode is the
default?).

To enable (partial) support for c++23, msvc needs '/std:c++latest'.
See here for a small experiment:
    https://godbolt.org/z/b6hM86Pxv