ImGuiConsole: fix scrolling through console history
Regression introduced in a7ff9027212b38.
But now that we upgraded to Dear ImGui v1.91.7 we can go back to our
original solution.

The underlying problem is was that 'ImGui::IsItemEdited()' also returns
'true' for up/down, while with the old callback approach we could
separate this wrong other edit actions.