Partly replace, partly rename our ranges::copy() function
We had 2 overloads of ranges::copy()
* One compatible with std::ranges::copy(), this has been replaced with
  the std version. It copies from a range to an output-iterator.
* One that takes another ranges (instead of output-iterator) as
  destination.

This 2nd version can do strictly less than the 1st one, but it does have
to advantage that (for sized-ranges) we can assert that the destination
range is at least as large as the source range. So I do want to keep
this version.

To avoid confusion between std::ranges::copy() and ranges::copy(), I
decided to rename our version (the 2nd overload) to copy_to_range().
This emphasises that the difference compared to the std version is the
destination parameter.

--

(For now) this finishes the replacement of our ranges::xxx() functions
with the corresponding std::ranges::xxx() functions.