Lazily create 'Poller' object in 'MidiInReader', 'RS232Net' and 'RS232Tester'
We recently found out that having multiple machine instances in openMSX
keeps a lot of file handles open. And opening ~100 machine instances can
run into the max-open-files-limit (default limit on my system is 1024).

Partly this is expected: the current implementation keeps an open file
handle for every system rom. This is because of how we handle mmap(), I
don't think it's strictly needed to keep the file handle open after
mmap(), but that's less convenient in the current code structure. And
this design still stems from way before we had multiple machine
instances. And is having ~100 machine instances really something that we
need?

But there was also an unexpected part: each machine instance also opened
6 file handles connected to a pipe. More specifically 3 pipes with both
end-points held by openMSX. These come from the 'Poller' object. This is
a utility that helps abort blocking read() calls (to be able to stop
helper threads(*), see comments in Poller for details).

The 'Poller' objects were located in 'MidiInReader', 'RS232Net' and
'RS232Tester'. These are 3 (rarely used?) pluggables. They (used to)
eagerly construct the Poller object, but it's actually only needed once
the Pluggable in plugged in. So this patch delays the construction till
we indeed plug in the Pluggable.

This is not a full solution, you'll still run into the
max-open-files-limit when using ~200 machine instances. But that's
probably already more than enough? And this 'solution' was reasonably
easy to implement.

(*) Using a helper thread (mostly) only to do a blocking read() is maybe
not the best idea. Though it is easy (especially when taking portability
to different operating systems into account). In the future we could try
to integrate a 3rdparty library that has an event loop with
file-read-ready events. But IMHO that's low priority.