Completer.hh: constrain iterator/range related template parameters
Use c++ concepts to limit (or constrain) the allowed types of how a
template can be instantiated.

For example:
    template<typename Range>
    void myFunction(Range&& range) { ... }
can be instantiated with any type. But probably it only makes sense to
instantiate this with types that actually model the concept of a range.

If we change the example to:
    template<std::ranges::forward_range Range>
    void myFunction(Range&& range) { ... }
then the compiler will check that the type with which we (try to)
instantiate indeed fits the requirements of a range, and more
specifically a "forward range" (see here for a list and description of
some range-related c++ concepts: https://en.cppreference.com/w/cpp/header/ranges)

This patch should have _no_ effect on the generated code. It only
documents the intent better, and it may help catch mistakes sooner
(rather than getting pages of compiler errors you hopefully get a
clearer more to the point compiler error).