Added initial version of game shuffler script.
Some streamers like to use that to create an entertaining stream... you point
the script to a folder with ROM files and it will run all the games in them,
time sliced. The interval is one of the options, as well as random or
round_robin order.
You can also change the machine or the extensions.

For now, only ROM files, to see if this catches on.

Thanks to Wouter for the code improvements and reviews.