Fix changing scale_factor when paused
See
  https://github.com/openMSX/openMSX/issues/1832

When the MSX VDP emulation finishes a frame, we upload raw-frame data as
textures. When the host needs to render (=scale) a frame we use that
data. Typically there's a one-to-one correspondence between these two
steps. Though not so during pause (or very slow emulation speed etc)
then we reuse the same texture data to render multiple host frames.

Together with the texture data we also split the screen into vertical
'regions'. Those correspond to screen-splits (more specifically between
screen modes with different horizontal resolution), but also changes
between vertical border and display area. This is needed because scaling
behaves different for such regions. The details of this split (also)
depend on the vertical output size. So we should recalculate this when
the vertical output resolution changes (so when scale_factor changes).